backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Django==5.2.8
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.3
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
fastapi==0.129.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
opencv-python-headless==4.13.0.92
packaging==25.0
pillow==12.1.1
polars-runtime-32==1.38.1
polars==1.38.1
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
starlette==0.52.1
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics-thop==2.0.18
ultralytics==8.4.14
urllib3==2.6.3
uvicorn==0.40.0
annotated-doc==0.0.4
